requirements.txt
anndata==0.12.2
matplotlib==3.10.5
numpy==2.3.2
pandas==2.3.1
scanpy==1.11.4
scikit_learn==1.7.1
torch==2.8.0
umap_learn==0.5.9.post2
